feat(type2): Skip jumpdest analysis (#631)
* Remove JDA for type2

* Perform sanity checks

* Clippy
