+++ evm_arithmetization/src/cpu/kernel/asm/core/exception.asm
@@ -124,22 +124,26 @@ global exc_invalid_jumpi_destination:
 
 
 global invalid_jump_jumpi_destination_common:
-    // We have a jump destination on the stack. We want to `PANIC` if it is valid, and jump to
-    // `fault_exception` if it is not. An address is a valid jump destination if it points to a
-    // `JUMPDEST` instruction. In practice, since in this implementation memory addresses are
-    // limited to 32 bits, we check two things:
-    //  1. the address is no more than 32 bits long, and
-    //  2. it points to a `JUMPDEST` instruction.
-    // stack: jump_dest
-    DUP1
-    %shr_const(32)
-    %jumpi(fault_exception) // This keeps one copy of jump_dest on the stack, but that's fine.
-    // jump_dest is a valid address; check if it points to a `JUMP_DEST`.
-    DUP1
-    %verify_non_jumpdest
-    %mload_current(@SEGMENT_JUMPDEST_BITS)
-    // stack: is_valid_jumpdest
-    %jumpi(panic) // Trap should never have been entered.
+    // We skip jump destinations verification with `cdk_erigon`.
+    #[cfg(not(feature = cdk_erigon))]
+    {
+        // We have a jump destination on the stack. We want to `PANIC` if it is valid, and jump to
+        // `fault_exception` if it is not. An address is a valid jump destination if it points to a
+        // `JUMPDEST` instruction. In practice, since in this implementation memory addresses are
+        // limited to 32 bits, we check two things:
+        //  1. the address is no more than 32 bits long, and
+        //  2. it points to a `JUMPDEST` instruction.
+        // stack: jump_dest
+        DUP1
+        %shr_const(32)
+        %jumpi(fault_exception) // This keeps one copy of jump_dest on the stack, but that's fine.
+        // jump_dest is a valid address; check if it points to a `JUMP_DEST`.
+        DUP1
+        %verify_non_jumpdest
+        %mload_current(@SEGMENT_JUMPDEST_BITS)
+        // stack: is_valid_jumpdest
+        %jumpi(panic) // Trap should never have been entered.
+    }
     %jump(fault_exception)
