+++ evm_arithmetization/src/cpu/kernel/tests/core/jumpdest_analysis.rs
@@ -1,13 +1,52 @@
+#![cfg(not(feature = "cdk_erigon"))]
+
 use std::collections::{BTreeSet, HashMap};
 
 use anyhow::Result;
 use ethereum_types::U256;
 use plonky2::field::goldilocks_field::GoldilocksField as F;
+use plonky2::hash::hash_types::RichField;
 
 use crate::cpu::kernel::aggregator::KERNEL;
 use crate::cpu::kernel::interpreter::Interpreter;
 use crate::cpu::kernel::opcodes::{get_opcode, get_push_opcode};
+use crate::memory::segments::Segment;
+use crate::witness::memory::MemoryAddress;
 use crate::witness::operation::CONTEXT_SCALING_FACTOR;
+
+impl<F: RichField> Interpreter<F> {
+    pub(crate) fn set_jumpdest_analysis_inputs(&mut self, jumps: HashMap<usize, BTreeSet<usize>>) {
+        self.generation_state.set_jumpdest_analysis_inputs(jumps);
+    }
+
+    pub(crate) fn get_jumpdest_bit(&self, offset: usize) -> U256 {
+        if self.generation_state.memory.contexts[self.context()].segments
+            [Segment::JumpdestBits.unscale()]
+        .content
+        .len()
+            > offset
+        {
+            // Even though we are in the interpreter, `JumpdestBits` is not part of the
+            // preinitialized segments, so we don't need to carry out the additional checks
+            // when get the value from memory.
+            self.generation_state.memory.get_with_init(MemoryAddress {
+                context: self.context(),
+                segment: Segment::JumpdestBits.unscale(),
+                virt: offset,
+            })
+        } else {
+            0.into()
+        }
+    }
+
+    pub(crate) fn get_jumpdest_bits(&self, context: usize) -> Vec<bool> {
+        self.generation_state.memory.contexts[context].segments[Segment::JumpdestBits.unscale()]
+            .content
+            .iter()
+            .map(|x| x.unwrap_or_default().bit(0))
+            .collect()
+    }
+}
 
 #[test]
 fn test_jumpdest_analysis() -> Result<()> {
